fix: :bug: Improve the logic to show statistics.
Improve the logic to show statistics. Now the admin generates statistics for all tasks or all a user if they pass an ID. In the case of users who are not admin, they only show statistics for the tasks they have associated with them., Improve the logic to show statistics. Now the admin generates statistics for all tasks or all a user if they pass an ID. In the case of users who are not admin, they only show statistics for the tasks they have associated.